Move stuff back to rpc-methods
