+++ packages/rpc-methods/src/__fixtures__/entropy.ts
