+++ packages/rpc-methods/src/restricted/getEntropy.ts
@@ -4,10 +4,12 @@ import {
   RestrictedMethodOptions,
   ValidPermissionSpecification,
 } from '@metamask/permission-controller';
-import { deriveEntropy, SIP_6_MAGIC_VALUE } from '@metamask/snaps-utils';
+import { SIP_6_MAGIC_VALUE } from '@metamask/snaps-utils';
 import { assertStruct, Hex, NonEmptyArray } from '@metamask/utils';
 import { ethErrors } from 'eth-rpc-errors';
 import { Infer, literal, object, optional, string } from 'superstruct';
+
+import { deriveEntropy } from '../utils';
 
 const targetKey = 'snap_getEntropy';
