+++ packages/rpc-methods/src/utils.test.ts
@@ -1,5 +1,7 @@
+import { SIP_6_MAGIC_VALUE } from '@metamask/snaps-utils';
+
 import { ENTROPY_VECTORS } from './__fixtures__';
-import { deriveEntropy, SIP_6_MAGIC_VALUE } from './entropy';
+import { deriveEntropy } from './utils';
 
 const TEST_SECRET_RECOVERY_PHRASE =
   'test test test test test test test test test test test ball';
